@gunnarmorling 2024-01-28
00:04.884 now. As for your question on the original PR, the official eval is on 8 cores indeed, but I am doing an occasional side eval on all the cores, with SMT, for funsies.